refactor: refactor product update page
