+++ src/modules/product-update/components/ProductUpdateForm.types.ts
@@ -1,0 +1,8 @@
+export type TProductUpdateFormFields = {
+  title: string;
+  categoryIds: string[];
+  description: string;
+  price: number;
+  rent: number;
+  rentOption: string;
+};
