+++ src/modules/product-update/components/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductUpdateForm";
