+++ src/modules/product-update/containers/ProductUpdateContainer.tsx
@@ -1,8 +1,8 @@
-import ProductUpdateForm from "../organisms/ProductUpdateForm";
+import ProductUpdateForm from "../components";
 
-const ProductUpdateLayout = () => {
+const ProductUpdateContainer = () => {
   return (
-    <div className="min-h-screen flex items-center justify-center">
+    <div className="flex items-center justify-center">
       <div className="w-4/5 md:w-3/5">
         <h1 className="text-3xl text-jet-black text-center mb-8">
           Edit Product Details
@@ -13,4 +13,4 @@ const ProductUpdateLayout = () => {
   );
 };
 
-export default ProductUpdateLayout;
+export default ProductUpdateContainer;
