+++ src/modules/product-update/containers/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductUpdateContainer";
