+++ src/shared/components/MultiSelect/MultiSelect.tsx
@@ -1,59 +1,26 @@
 import { useState, useRef, useEffect } from "react";
-import Button from "../atoms/Button";
-import CloseIcon from "@/assets/icons/close-icon.svg";
-import UnfoldMoreIcon from "@/assets/icons/unfold-more-icon.svg";
+import { X, ChevronsUpDown } from "lucide-react";
+import { Button } from "../shadui/button";
+import { TMultiSelectProps, TOption } from "./MultiSelect.types";
 
-type Option = {
-  id: string;
-  name: string;
-};
-
-type Props = {
-  placeholder: string;
-  options: Option[];
-  onChange: (selected: Option[]) => void;
-  error?: boolean;
-  defaultSelected?: Option[];
-};
-
-const MultiSelect = (props: Props) => {
-  const [selectedOptions, setSelectedOptions] = useState<Option[]>(
-    props.defaultSelected || []
-  );
+const MultiSelect = (props: TMultiSelectProps) => {
+  const [selectedOptions, setSelectedOptions] = useState<TOption[]>(props.defaultSelected || []);
   const [isOpen, setIsOpen] = useState(false);
   const dropdownRef = useRef<HTMLDivElement | null>(null);
+  const hasInitialized = useRef(false);
+
+  const { defaultSelected, onChange } = props;
 
   useEffect(() => {
-    if (props.defaultSelected?.length) {
-      props.onChange(props.defaultSelected);
+    if (!hasInitialized.current && defaultSelected?.length) {
+      onChange(defaultSelected);
+      hasInitialized.current = true;
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const toggleDropdown = () => setIsOpen(!isOpen);
-
-  const selectOption = (option: Option) => {
-    if (!selectedOptions.some((selected) => selected.id === option.id)) {
-      const newSelectedOptions = [...selectedOptions, option];
-      setSelectedOptions(newSelectedOptions);
-      props.onChange(newSelectedOptions);
-    }
-  };
-
-  const removeOption = (option: Option) => {
-    const newSelectedOptions = selectedOptions.filter(
-      (prevOption) => prevOption.id !== option.id
-    );
-    setSelectedOptions(newSelectedOptions);
-    props.onChange(newSelectedOptions);
-  };
+  }, [defaultSelected, onChange]);
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
-      if (
-        dropdownRef.current &&
-        !dropdownRef.current.contains(event.target as Node)
-      ) {
+      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
         setIsOpen(false);
       }
     };
@@ -63,15 +30,29 @@ const MultiSelect = (props: Props) => {
     };
   }, []);
 
+  const selectOption = (option: TOption) => {
+    if (!selectedOptions.some((selected) => selected.id === option.id)) {
+      const newSelectedOptions = [...selectedOptions, option];
+      setSelectedOptions(newSelectedOptions);
+      props.onChange(newSelectedOptions);
+    }
+  };
+
+  const removeOption = (option: TOption) => {
+    const newSelectedOptions = selectedOptions.filter((prevOption) => prevOption.id !== option.id);
+    setSelectedOptions(newSelectedOptions);
+    props.onChange(newSelectedOptions);
+  };
+
   return (
     <div className="relative w-full" ref={dropdownRef}>
       <div
         className={`rounded-sm cursor-pointer flex justify-between items-center overflow-auto ${
           props.error
             ? `border border-red-500 focus:outline-red-500`
-            : `focus:outline-purple border-2 border-gray`
+            : `focus:outline-purple border border-input`
         }`}
-        onClick={toggleDropdown}
+        onClick={() => setIsOpen(!isOpen)}
       >
         <div className="flex gap-1">
           {selectedOptions.length === 0 ? (
@@ -80,24 +61,24 @@ const MultiSelect = (props: Props) => {
             selectedOptions.map((option) => (
               <span
                 key={option.id}
-                className="flex items-center gap-1 border-2 border-gray px-2 py-1.5 m-1"
+                className="flex items-center gap-1 border border-input px-2 py-1.5 m-1"
               >
                 {option.name}
                 <Button
-                  className="cursor-pointer"
+                  variant={"ghost"}
                   onClick={(e) => {
                     e.stopPropagation();
                     removeOption(option);
                   }}
                 >
-                  <CloseIcon />
+                  <X />
                 </Button>
               </span>
             ))
           )}
         </div>
         <span className="pr-2">
-          <UnfoldMoreIcon />
+          <ChevronsUpDown size={14} />
         </span>
       </div>
       {isOpen && (
