+++ src/shared/components/MultiSelect/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MultiSelect";
