+++ src/shared/components/Select/Select.tsx
@@ -5,7 +5,7 @@ type Props = {
 
 const Select = (props: Props) => {
   return (
-    <select className="border-2 border-gray rounded-sm outline-gray p-3 w-full" {...props.field}>
+    <select className="border border-input rounded-sm outline-gray p-3 w-full" {...props.field}>
       {props.options.map((option, index) => (
         <option key={index} value={option.value}>
           {option.label}
